# src/main/resources/static/docs/1.0.0/swagger/todo-1.0.0-swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "API de lista de afazeres\n",
    "version" : "1.0.0",
    "title" : "Todo list API",
    "termsOfService" : "http://swagger.io/terms/",
    "contact" : {
      "name" : "Lucas Bovolini",
      "email" : "lbovolini94@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers" : [ {
    "description" : "SwaggerHub API Auto Mocking",
    "url" : "https://virtserver.swaggerhub.com/lbovolini/Todo/1.0.0"
  }, {
    "description" : "localhost",
    "url" : "http://localhost:8080/todo"
  } ],
...
  "paths" : {
...
    "/api/v1/users" : {
      "put" : {
        "tags" : [ "User API" ],
        "summary" : "Atualiza todos os atributos do usuário na base de dados",
        "operationId" : "update",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              },
              "example" : {
                "id" : "5ff46c3c9ad0cb476eff6a5d",
                "username" : "lucas",
                "password" : "lucas1123"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Operação realizada com sucesso",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                },
                "example" : {
                  "id" : "5ff46c3c9ad0cb476eff6a5d",
                  "username" : "lucas"
                }
              }
            }
          },
          "400" : {
            "description" : "Argumento inválido",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "timestamp" : "2021-01-07T17:07:21.817070849+00:00",
                  "status" : 400,
                  "error" : "Bad Request",
                  "message" : "Method Argument Not Valid",
                  "path" : "/todo/api/v1/users/",
                  "errors" : [ {
                    "field" : "password",
                    "message" : "must not be blank",
                    "value" : null
                  } ]
                }
              }
            }
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "description" : "Usuário não encontrado",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          }
        },
        "security" : [ {
          "todoAuth" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "User API" ],
        "summary" : "Salva o usuário na base de dados",
        "operationId" : "save",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              },
              "example" : {
                "username" : "lucas",
                "password" : "lucas1123"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Operação realizada com sucesso",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                },
                "example" : {
                  "id" : "5ff742ba4722e0108945a837",
                  "username" : "lbovolini"
                }
              }
            }
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "409" : {
            "description" : "A base de dados já possui um usuário com pelo menos um dos campos únicos informados",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "timestamp" : "2021-01-07T17:19:10.474787777+00:00",
                  "status" : 409,
                  "error" : "Conflict",
                  "message" : "Duplicate Key",
                  "path" : "/todo/api/v1/users/",
                  "errors" : [ {
                    "field" : "username",
                    "message" : "Duplicate Key",
                    "value" : "lucas"
                  } ]
                }
              }
            }
          }
        },
        "security" : [ {
          "todoAuth" : [ ]
        } ]
      }
    },
...
    "/api/v1/users/{id}" : {
      "get" : {
        "tags" : [ "User API" ],
        "summary" : "Realiza a busca do usuário pelo id informado",
        "operationId" : "findById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "5ff46c3c9ad0cb476eff6a5d",
          "description" : "Id do usuário registrado na base de dados que será pesquisado"
        } ],
        "responses" : {
          "200" : {
            "description" : "Operação realizada com sucesso",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                },
                "example" : {
                  "id" : "5fecae92fc4878605c5c7a5b",
                  "username" : "lucas"
                }
              }
            }
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "description" : "Usuário com o id informado não encontrado",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          }
        },
        "security" : [ {
          "todoAuth" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "User API" ],
        "summary" : "Remove o usuário com o id informado",
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "5ff46c3c9ad0cb476eff6a5d",
          "description" : "Id do usuário registrado na base de dados que será removido"
        } ],
        "responses" : {
          "204" : {
            "description" : "Usuário removido com sucesso",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "description" : "Usuário não encontrado",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "timestamp" : "2021-01-07T21:22:46.433726467+00:00",
                  "status" : 404,
                  "error" : "Not Found",
                  "message" : "User not found",
                  "path" : "/todo/api/v1/users/11"
                }
              }
            }
          }
        },
        "security" : [ {
          "todoAuth" : [ ]
        } ]
      }
    },
    "/api/v1/todos/{id}" : {
      "get" : {
        "tags" : [ "Todo API" ],
        "summary" : "Realiza a busca da lista de afazeres pelo id informado",
        "operationId" : "findById_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "5ff65c7440deeb6df754110f",
          "description" : "Id da lista de afazeres registrada na base de dados que será pesquisada"
        } ],
        "responses" : {
          "200" : {
            "description" : "Operação realizada com sucesso",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Todo"
                },
                "example" : {
                  "id" : "5ff65c7440deeb6df754110f",
                  "userId" : "5ff4a5f993002c22433de9e3",
                  "tasks" : [ {
                    "name" : "Teste",
                    "description" : "Enviar teste",
                    "date" : "2021-02-04T17:47:25.147613061+00:00"
                  } ]
                }
              }
            }
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "description" : "Lista de afazeres com o id informado não encontrada",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          }
        },
        "security" : [ {
          "todoAuth" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Todo API" ],
        "summary" : "Remove a lista de afazeres correspondente ao id informado",
        "operationId" : "delete_1",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          },
          "example" : "5ff65c7440deeb6df754110f",
          "description" : "Id da lista de afazeres registrada na base de dados que será removida"
        } ],
        "responses" : {
          "204" : {
            "description" : "Lista de afazeres removida com sucesso",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "description" : "Lista de afazeres com o id informado não encontrada",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "timestamp" : "2021-01-06T22:25:52.641697330+00:00",
                  "status" : 404,
                  "error" : "Not Found",
                  "message" : "Todo not found",
                  "path" : "/todo/api/v1/todos/1"
                }
              }
            }
          }
        },
        "security" : [ {
          "todoAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "responses" : {
      "Unauthorized" : {
        "description" : "Usuário não autenticado",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            },
            "example" : {
              "timestamp" : "2020-12-31T17:31:04.518+00:00",
              "status" : 401,
              "error" : "Unauthorized",
              "message" : "Invalid token",
              "path" : "..."
            }
          }
        }
      },
      "Forbidden" : {
        "description" : "Usuário não possui permissão para acessar este recurso ou serviço",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            },
            "example" : {
              "timestamp" : "2021-01-11T21:12:27.253+00:00",
              "status" : 403,
              "error" : "Forbidden",
              "message" : "",
              "path" : "..."
            }
          }
        }
      }
    },
    "schemas" : {
      "User" : {
        "description" : "Usuário do sistema",
        "required" : [ "username", "password" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "description" : "Identificador único",
            "type" : "string"
          },
          "username" : {
            "description" : "Nome de usuário. Deve conter no mínimo 3 caracteres e no máximo 20, sendo estes letras ou números",
            "pattern" : "^(\\w){3,20}\\b",
            "type" : "string"
          },
          "password" : {
            "description" : "Senha. Deve conter no mínimo 8 caracteres, com no mínimo 1 letra e 1 número",
            "pattern" : "^(?=.*[\\d])(?=.*[a-z])[\\w!@#$%^&*()-=+,.;:]{8,}$",
            "type" : "string",
            "writeOnly" : true
          },
          "authorities" : {
            "description" : "Lista de permissões do usuário",
            "type" : "array",
            "items" : {
              "type" : "object"
            },
            "readOnly" : true,
            "writeOnly" : true
          },
          "accountNonLocked" : {
            "description" : "Indica se a conta do usuário está bloqueada ou não. A conta é sempre bloqueada a cada logout e desbloqueada a cada login",
            "type" : "boolean",
            "readOnly" : true,
            "writeOnly" : true
          }
        }
      },
...
      "Todo" : {
        "required" : [ "userId" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "description" : "Identificador único",
            "type" : "string"
          },
          "userId" : {
            "description" : "Identificador único do usuário a quem pertence esta lista de afazeres",
            "type" : "string"
          },
          "tasks" : {
            "description" : "Lista de tarefas",
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Task"
            }
          },
          "attachments" : {
            "description" : "Lista de anexos",
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Attachment"
            }
          }
        }
      },
      "Task" : {
        "required" : [ "date", "description", "name" ],
        "type" : "object",
        "properties" : {
          "name" : {
            "description" : "Nome",
            "type" : "string"
          },
          "description" : {
            "description" : "Descrição",
            "type" : "string"
          },
          "date" : {
            "description" : "Data",
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "Attachment" : {
        "required" : [ "path" ],
        "type" : "object",
        "properties" : {
          "name" : {
            "description" : "Nome do arquivo",
            "type" : "string",
            "readOnly" : true
          },
          "path" : {
            "description" : "Caminho não absoluto do arquivo",
            "type" : "string"
          },
          "type" : {
            "description" : "Tipo do arquivo",
            "type" : "string",
            "readOnly" : true
          },
          "extension" : {
            "description" : "Extensão do arquivo",
            "type" : "string",
            "readOnly" : true
          },
          "size" : {
            "description" : "Tamanho do arquivo",
            "type" : "integer",
            "format" : "int64",
            "readOnly" : true
          }
        }
      },
      "ApiError" : {
        "required" : [ "timestamp", "status", "error", "message", "path" ],
        "type" : "object",
        "properties" : {
          "timestamp" : {
            "description" : "Momento em que ocorreu o erro",
            "type" : "string"
          },
          "status" : {
            "description" : "Status HTTP do erro",
            "type" : "integer"
          },
          "error" : {
            "description" : "Mensagem de erro do status HTTP",
            "type" : "string"
          },
          "message" : {
            "description" : "Mensagem de erro lançada pela exceção, ou mensagem personalizada",
            "type" : "string"
          },
          "path" : {
            "description" : "Caminho relativo do endpoint da solicitação em que ocorreu o erro",
            "type" : "string"
          },
          "errors" : {
            "description" : "Lista de erros",
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Error"
            }
          }
        }
      },
      "Error" : {
        "required" : [ "field", "message", "value" ],
        "type" : "object",
        "properties" : {
          "field" : {
            "description" : "Nome do campo em que ocorreu o erro",
            "type" : "string"
          },
          "message" : {
            "description" : "Mensagem de erro",
            "type" : "string"
          },
          "value" : {
            "description" : "Valor informado que gerou o erro",
            "type" : "string"
          }
        }
      }
    },
    "securitySchemes" : {
      "todoAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}